+++ release-notes.md
@@ -7,6 +7,7 @@ Changes:
 
 Fixes:
   - Prevent using invalid characters when adding labels to a study.
+  - [Issue #58](https://github.com/orthanc-server/orthanc-explorer-2/issues/58) Upload function only uploads the first file.
 
 1.4.0 (2024-05-16)
 ==================